Настройка роли доступа

+++ src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
@@ -5,12 +5,14 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import ru.skypro.homework.dto.*;
+import ru.skypro.homework.dto.AdsDto;
+import ru.skypro.homework.dto.CreateAdsDto;
+import ru.skypro.homework.dto.FullAdsDto;
+import ru.skypro.homework.dto.ResponseWrapperAds;
 import ru.skypro.homework.exception.AdsNotFoundException;
 import ru.skypro.homework.mapper.AdsMapper;
 import ru.skypro.homework.mapper.UserMapper;
 import ru.skypro.homework.model.Ads;
-import ru.skypro.homework.model.Comment;
 import ru.skypro.homework.model.Image;
 import ru.skypro.homework.model.User;
 import ru.skypro.homework.repository.AdsRepository;
@@ -18,11 +20,10 @@ import ru.skypro.homework.repository.CommentRepository;
 import ru.skypro.homework.repository.UserRepository;
 import ru.skypro.homework.service.AdsService;
 import ru.skypro.homework.service.ImageService;
+import ru.skypro.homework.service.SecurityService;
 import ru.skypro.homework.service.UserService;
 
 import java.util.List;
-
-import static org.springframework.util.ObjectUtils.isEmpty;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -34,9 +35,10 @@ public class AdsServiceImpl implements AdsService {
     private final UserMapper userMapper;
     private final ImageService imageService;
     private final UserService userService;
-//    private final CommentService commentService;
+    //    private final CommentService commentService;
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
+    private final SecurityService securityService;
 
     @Override
     public ResponseWrapperAds getAllAds() {
@@ -49,15 +51,14 @@ public class AdsServiceImpl implements AdsService {
 
     @Override
     public AdsDto createAds(CreateAdsDto createAdsDto, MultipartFile image, Authentication authentication) {
-        UserDto userDto = userService.getUserByEmail(authentication.getName());
+        User currentUser = userService.getUser(authentication.getName());
 
         Ads ads = adsMapper.INSTANCE.createAdsDtoToAds(createAdsDto);
-        ads.setUser(userMapper.INSTANCE.userDtoToUser(userDto));
+        ads.setUser(currentUser);
         Ads savedAds = adsRepository.save(ads);
 
-        Image adsImage = imageService.createImage(image, savedAds); // а если неск фото
-        adsImage.setAds(savedAds);
-        adsImage.setUser(userMapper.INSTANCE.userDtoToUser(userDto));
+        Image adsImage = imageService.createImage(image, savedAds);
+        savedAds.setImages(adsImage);
 
         return adsMapper.INSTANCE.adsToAdsDto(savedAds);
     }
@@ -75,15 +76,14 @@ public class AdsServiceImpl implements AdsService {
 
     @Override
     public void removeAds(Integer id, Authentication authentication) {
-        Ads ads = getAdsById(id);
+//        Ads ads = getAdsById(id);
 
 //        checkIfUserCanAlterAds(authentication, ads); // доработать метод проверки
-
-        List<Comment> comments = ads.getComments();
-        comments.stream()
-                .forEach(comment -> commentRepository.deleteById(comment.getId()));
-
-        adsRepository.deleteById(id);
+//        List<Comment> comments = ads.getComments();
+//        comments.stream()
+//                .forEach(comment -> commentRepository.deleteById(comment.getId()));
+        if (securityService.accessAds(authentication, id))
+            adsRepository.deleteById(id);
     }
 
     @Override
@@ -91,16 +91,16 @@ public class AdsServiceImpl implements AdsService {
         Ads oldAds = getAdsById(id);
 
 //        checkIfUserCanAlterAds(authentication, oldAds); // доработать метод проверки
-
-        Ads infoToUpdate = adsMapper.INSTANCE.createAdsDtoToAds(createAdsDto);
-
-        oldAds.setPrice(infoToUpdate.getPrice());
-        oldAds.setTitle(infoToUpdate.getTitle());
-        oldAds.setDescription(infoToUpdate.getDescription());
-
+        if (securityService.accessAds(authentication, id)) {
+            Ads infoToUpdate = adsMapper.INSTANCE.createAdsDtoToAds(createAdsDto);
+            oldAds.setPrice(infoToUpdate.getPrice());
+            oldAds.setTitle(infoToUpdate.getTitle());
+            oldAds.setDescription(infoToUpdate.getDescription());
+        }
         Ads updatedAds = adsRepository.save(oldAds);
         return adsMapper.INSTANCE.adsToAdsDto(updatedAds);
     }
+
 
     @Override
     public ResponseWrapperAds getAllAdsForUser(String username) {
@@ -123,8 +123,8 @@ public class AdsServiceImpl implements AdsService {
         return adsMapper.INSTANCE.adsListToResponseWrapperAds(adsDtoDtoList.size(), adsDtoDtoList);
     }
 
-    private void checkIfUserCanAlterAds(Authentication authentication, Ads ads){
-        if(ads.getUser().getEmail() != authentication.getName()){
+    private void checkIfUserCanAlterAds(Authentication authentication, Ads ads) {
+        if (ads.getUser().getEmail() != authentication.getName()) {
             throw new RuntimeException("Вы не имеете права доступа");
         }
     }
