+++ src/main/java/ru/skypro/homework/service/impl/CommentServiceImpl.java
@@ -2,15 +2,11 @@ package ru.skypro.homework.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.LoggerFactory;
-import org.slf4j.Logger;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import ru.skypro.homework.dto.CommentDto;
 import ru.skypro.homework.dto.ResponseWrapperComment;
-import ru.skypro.homework.dto.UserDto;
 import ru.skypro.homework.exception.CommentNotFoundException;
-import ru.skypro.homework.exception.CommentsNotFoundException;
 import ru.skypro.homework.mapper.CommentMapper;
 import ru.skypro.homework.mapper.UserMapper;
 import ru.skypro.homework.model.Ads;
@@ -19,10 +15,10 @@ import ru.skypro.homework.model.User;
 import ru.skypro.homework.repository.CommentRepository;
 import ru.skypro.homework.service.AdsService;
 import ru.skypro.homework.service.CommentService;
+import ru.skypro.homework.service.SecurityService;
 import ru.skypro.homework.service.UserService;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 
 import static ru.skypro.homework.mapper.CommentMapper.INSTANCE;
@@ -38,8 +34,9 @@ public class CommentServiceImpl implements CommentService {
     private final UserMapper userMapper;
     private final CommentRepository commentRepository;
     private final UserService userService;
+    private final SecurityService securityService;
 
-    private final Logger logger = LoggerFactory.getLogger(CommentServiceImpl.class);
+//    private final Logger logger = LoggerFactory.getLogger(CommentServiceImpl.class);
 
 
     @Override
@@ -52,9 +49,11 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public CommentDto createNewComment(Integer adsId, CommentDto commentDto, Authentication authentication) {
+        log.info("Was invoked createNewComment method from {}", CommentService.class.getSimpleName());
         Ads ads = adsService.getAdsById(adsId); // возможно исключение AdsNotFoundException
-        UserDto userDto = userService.getUserByEmail(authentication.getName());
-        User currentUser = userMapper.userDtoToUser(userDto);
+//        UserDto userDto = userService.getUserByEmail(authentication.getName());
+//        User currentUser = userMapper.userDtoToUser(userDto);
+        User currentUser = userService.getUser(authentication.getName());
 
         Comment comment = INSTANCE.commentDtoToComment(commentDto);
         comment.setAds(ads);
@@ -66,28 +65,29 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public CommentDto getComment(Integer adPk, Integer id) {
-        Comment comment = commentRepository.findAdsComment(adPk,id).orElseThrow(CommentNotFoundException::new);
+        Comment comment = commentRepository.findAdsComment(adPk, id).orElseThrow(CommentNotFoundException::new);
         return INSTANCE.commentToCommentDto(comment);
     }
 
     @Override
     public void deleteComment(Integer adPk, Integer id, Authentication authentication) {
-        Comment comment = commentRepository.findAdsComment(adPk,id).orElseThrow(CommentNotFoundException::new);
+        Comment comment = commentRepository.findAdsComment(adPk, id).orElseThrow(CommentNotFoundException::new);
 //        checkIfUserCanAlterComment(authentication, comment); // доработать метод проверки
-
-        commentRepository.delete(comment);
+        if (securityService.accessComments(authentication, id)) {
+            commentRepository.delete(comment);
+        }
     }
 
     @Override
     public CommentDto updateComment(Integer adPk, Integer id, CommentDto commentDto, Authentication authentication) {
-        Comment comment = commentRepository.findAdsComment(adPk,id)
+        Comment comment = commentRepository.findAdsComment(adPk, id)
                 .orElseThrow(CommentNotFoundException::new);
-
 //        checkIfUserCanAlterComment(authentication, comment); // доработать метод проверки
-        comment.setText(commentDto.getText());
-        comment.setCreateAt(LocalDate.parse(commentDto.getCreateAt()));
-        commentRepository.save(comment);
-
+        if (securityService.accessComments(authentication, id)) {
+            comment.setText(commentDto.getText());
+            comment.setCreateAt(LocalDate.parse(commentDto.getCreateAt()));
+            commentRepository.save(comment);
+        }
         return INSTANCE.commentToCommentDto(comment);
 
     }
@@ -108,8 +108,8 @@ public class CommentServiceImpl implements CommentService {
 //        commentRepository.deleteById(id);
 //    }
 
-    private void checkIfUserCanAlterComment(Authentication authentication, Comment comment){
-        if(comment.getUser().getEmail() != authentication.getName()){
+    private void checkIfUserCanAlterComment(Authentication authentication, Comment comment) {
+        if (comment.getUser().getEmail() != authentication.getName()) {
             throw new RuntimeException("Вы не имеете права доступа");
         }
     }
